requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Werkzeug
wtforms
email_validator
Flask-WTF
Pillow
Flask-OAuth
Flask-OAuthlib
lazy
requests
requests-OAuthlib
SQLAlchemy
Jinja2
Flask-Mail
itsdangerous-2.2.0
python-dotenv